Updated get-process-function
